optimizing and fixing

+++ setup.py
@@ -33,22 +33,29 @@ extensions = [
 ]
 
 # Setup
-print("Generating stub files ...")
-os.makedirs(stubs_dir, exist_ok=True)
+GEN_STUBS = True
 for filename in os.listdir(source_dir):
-    if filename.endswith('.py'):
-        source_path = os.path.join(source_dir, filename)
-        subprocess.run(['stubgen', '-o', stubs_dir, source_path], check=True)
-for stub_file in os.listdir(os.path.join(stubs_dir, source_dir)):
-    if stub_file.endswith('.pyi'):
-        source_stub_path = os.path.join(stubs_dir, source_dir, stub_file)
-        if os.path.exists(os.path.join(source_dir, stub_file)):
-            print(f"Skipped moving {source_stub_path} because {os.path.join(source_dir, stub_file)} already exists!")
-        else:
-            shutil.move(source_stub_path, source_dir)
-            print(f"Moved {source_stub_path} to {source_dir}!")
-shutil.rmtree(os.path.join(stubs_dir))
-print("Stub files generated and moved successfully.")
+    if filename.endswith('.pyi'):
+        GEN_STUBS = False
+if GEN_STUBS is False:
+    print("Skipping stub files ...")
+else:
+    print("Generating stub files ...")
+    os.makedirs(stubs_dir, exist_ok=True)
+    for filename in os.listdir(source_dir):
+        if filename.endswith('.py'):
+            source_path = os.path.join(source_dir, filename)
+            subprocess.run(['stubgen', '-o', stubs_dir, source_path], check=True)
+    for stub_file in os.listdir(os.path.join(stubs_dir, source_dir)):
+        if stub_file.endswith('.pyi'):
+            source_stub_path = os.path.join(stubs_dir, source_dir, stub_file)
+            if os.path.exists(os.path.join(source_dir, stub_file)):
+                print(f"Skipped moving {source_stub_path} because {os.path.join(source_dir, stub_file)} already exists!")
+            else:
+                shutil.move(source_stub_path, source_dir)
+                print(f"Moved {source_stub_path} to {source_dir}!")
+    shutil.rmtree(os.path.join(stubs_dir))
+    print("Stub files generated and moved successfully.")
 
 with open("README.md", "r") as fh:
     print("Using README.md content as `long_description` ...")
@@ -65,7 +72,7 @@ setup(
      long_description=long_description,
      long_description_content_type="text/markdown",
      license='LSOSL - LUCIT Synergetic Open Source License',
-     install_requires=['lucit-licensing-python>=1.8.2', 'Cython', 'requests>=2.31.0',
+     install_requires=['lucit-licensing-python>=1.8.2', 'Cython>=3.0.10', 'requests>=2.31.0',
                        'unicorn-binance-websocket-api>=2.8.0', 'unicorn-binance-rest-api>=2.6.1'],
      keywords='binance, depth cache',
      project_urls={
